pyproject.toml
[project]
name = "data-game"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "mysql-connector-python>=9.3.0",
    "pandas>=2.3.0",
    "pipreqs>=0.5.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "scikit-learn>=1.7.0",
    "sqlalchemy>=2.0.41",
]
